+++ src/engine/renderer/vulkan/renderer_vulkan.cc
@@ -525,8 +525,8 @@ void RendererVulkan::DestroyGeometry(uint64_t resource_id) {
 }
 
 void RendererVulkan::Draw(uint64_t resource_id,
-                          uint64_t num_indices,
-                          uint64_t start_offset) {
+                          size_t num_indices,
+                          size_t start_offset) {
   auto it = geometries_.find(resource_id);
   if (it == geometries_.end())
     return;
@@ -629,7 +629,7 @@ void RendererVulkan::DestroyTexture(uint64_t resource_id) {
 }
 
 void RendererVulkan::ActivateTexture(uint64_t resource_id,
-                                     uint64_t texture_unit) {
+                                     size_t texture_unit) {
   auto it = textures_.find(resource_id);
   if (it == textures_.end())
     return;
